allow host programs to be run from another DOS program
